Fixed broken modularity and default stock removal disabled
