+++ app/code/Magento/Inventory/view/adminhtml/ui_component/inventory_stock_form.xml
@@ -77,7 +77,7 @@
                     <item name="breakLine" xsi:type="boolean">false</item>
                 </item>
             </argument>
-            <container class="Magento\Inventory\Ui\Component\AssignSourcesButton" name="assign_sources_button" component="Magento_Ui/js/form/components/button" template="ui/form/components/button/container">
+            <button name="assign_sources_button" template="ui/form/components/button/container">
                 <argument name="data" xsi:type="array">
                     <item name="config" xsi:type="array">
                         <item name="formElement" xsi:type="string">container</item>
@@ -93,7 +93,7 @@
                         <item name="dataScope" xsi:type="string"/>
                     </item>
                 </argument>
-            </container>
+            </button>
         </container>
         <dynamicRows name="assigned_sources" component="Magento_Ui/js/dynamic-rows/dynamic-rows-grid" sortOrder="20">
             <argument name="data" xsi:type="array">
