+++ app/code/Magento/InventoryCatalog/view/adminhtml/web/template/assign-sources-button.html
@@ -1,0 +1,19 @@
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<div class="admin__field admin__field-group-additional" css="$data.additionalClasses" visible="visible">
+    <label class="admin__field-label" if="$data.label" visible="$data.labelVisible">
+        <span text="label"/>
+    </label>
+
+    <div class="admin__field-control">
+        <render args="elementTmpl"/>
+    </div>
+
+    <div class="admin__field-note" if="$data.notice">
+        <span html="notice"/>
+    </div>
+</div>
